% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2value.R
\name{s2value}
\alias{s2value}
\title{A function for finding the z values.}
\usage{
s2value(z)
}
\arguments{
\item{z}{The z values for a dataset}
}
\value{
The s^2 value
}
\description{
A function for finding the z values.
}
\examples{
(enter z array, return standard value squared)
}
