% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zvalue.R
\name{sdvalue}
\alias{sdvalue}
\title{A function for finding the z values.}
\usage{
sdvalue(z)
}
\arguments{
\item{z}{The z values for a dataset}
}
\value{
The s^2 value
}
\description{
A function for finding the z values.
}
\examples{

}
